requirements.txt
# Email Attachment Processor Dependencies

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
# Uncomment as needed for your specific use case

# For email processing (IMAP/SMTP)
imaplib2>=3.6
email-validator>=2.0.0

# For file parsing and processing
PyPDF2>=3.0.0          # PDF parsing
openpyxl>=3.1.0        # Excel file processing
python-docx>=0.8.11    # Word document processing
pandas>=2.0.0          # CSV and data processing
Pillow>=10.0.0         # Image processing

# For data storage
google-cloud-storage>=2.10.0  # Google Cloud Storage

# For notifications

sendgrid>=6.10.0       # SendGrid email notifications

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0

# Optional but recommended
python-magic>=0.4.27     # file type detection
PyPDF2>=3.0.0
# slack-sdk>=3.21.0      # Slack notifications
# twilio>=8.5.0          # SMS notifications

# For logging and monitoring
# structlog>=23.1.0      # Structured logging
# prometheus-client>=0.17.0  # Prometheus metrics

# For testing
# pytest>=7.4.0
# pytest-mock>=3.11.0
# pytest-cov>=4.1.0

# For development
# black>=23.7.0          # Code formatting
# flake8>=6.0.0          # Linting
# mypy>=1.5.0            # Type checking
